add handle onclick utility component and update imports
