+++ factsheet/frontend/src/components/comparisonBoardItems.js
@@ -5,7 +5,8 @@ import Chip from '@mui/material/Chip';
 import StudyChip from '../styles/oep-theme/components/studyChip';
 import palette from '../styles/oep-theme/palette';
 import variables from '../styles/oep-theme/variables';
-import StudyKeywords from './scenarioBundleMicroComponents/StudyDescriptors.js';
+import StudyKeywords from './scenarioBundleUtilityComponents/StudyDescriptors.js';
+import handleOpenURL from './scenarioBundleUtilityComponents/handleOnClickTableIRI.js';
 
 const reorder = (list, startIndex, endIndex) => {
   const result = Array.from(list);
@@ -64,38 +65,6 @@ export default function  ComparisonBoardItems (props) {
       items: newItems,
     });
   }
-
-// enhance open url function to handle urls better
-// TODO change once OEKG is migrated
-const handleOpenURL = (url, setError) => {
-  if (!url[1]) {
-    setError('Invalid URL');
-    return;
-  }
-
-  let processedURL = url.trim();
-
-  // Handle URLs with spaces in the middle
-  processedURL = processedURL.replace(/\s+/g, '%20');
-
-  // Prepend base URL if not starting with http:// or https://
-  if (!processedURL.startsWith('http://') && !processedURL.startsWith('https://')) {
-    const baseURL = window.location.origin;
-    processedURL = `${baseURL}${processedURL}`;
-  }
-
-  if (processedURL.startsWith('http://') && !processedURL.startsWith('http://127') && !processedURL.startsWith('http://local')){
-    processedURL = processedURL.replace('http://', 'https://');
-  }
-
-  // Encode the URL to handle special characters
-  try {
-    const encodedURL = encodeURI(processedURL);
-    window.open(encodedURL, "_blank");
-  } catch (error) {
-    setError('Invalid URL format');
-  }
-};
 
   const label = { inputProps: { 'aria-label': 'Checkbox demo' } };
